microblog_bench/worker/requirements.txt
pkg-resources==0.0.0
PyYAML==5.3
six==1.14.0
thrift==0.13.0
